Extension docs: Start with general principles
I want this to start with the most simple example possible. And since
the introduction of flarum/core#1559 (per-site extenders), this does not
even require explaining extensions.

After that, I'd like to continue with the most simple extensions
possible (everything in one extender file), before moving on to where to
put classes if you want to structure your extension code as it gets
larger.

Refs flarum/core#1000.