Add Composer dependencies
Since it comes up so often and Composer is currently required for
installing Flarum, we may as well mention these as dependencies.

These do not need to be checked in the installer's prerequisites
check, as Composer will already have been run at that time. (In
other words, the installer only checks runtime dependencies.)

Closes #55.